Export reference asset XLRs from the react plugin
