+++ plugins/reference-assets/core/src/plugin.ts
@@ -1,28 +1,11 @@
-import type {
-  Player,
-  PlayerPlugin,
-  ExtendedPlayerPlugin,
-} from '@player-ui/player';
+import type { Player, PlayerPlugin } from '@player-ui/player';
 import { AssetTransformPlugin } from '@player-ui/asset-transform-plugin';
-import type {
-  ActionAsset,
-  CollectionAsset,
-  InfoAsset,
-  InputAsset,
-  TextAsset,
-} from './assets';
 import { inputTransform, actionTransform } from './assets';
 
 /**
  * A plugin to add transforms for the reference assets
  */
-export class ReferenceAssetsPlugin
-  implements
-    PlayerPlugin,
-    ExtendedPlayerPlugin<
-      [InputAsset, TextAsset, ActionAsset, InfoAsset, CollectionAsset]
-    >
-{
+export class ReferenceAssetsPlugin implements PlayerPlugin {
   name = 'reference-assets-transforms';
 
   apply(player: Player) {
